Add dual reed board support, extend button shiftregister to 24, translate docs to English

+++ code/src/controls/reedControl.h
@@ -4,7 +4,7 @@
 #include <Arduino.h>
 
 namespace ReedControl {
-    int read();
+    long read();
     void push(int buttonNr, void (*callback)());
     void updateStates();
     void setStates();
